Use float to prevent casting.
Former-commit-id: 4d0f0912927e86fb3dda26bc8131c2b61e4913d6
Former-commit-id: 87fc5322e14aa23b36a5edc91e2a539be87e6568
Former-commit-id: ba498a73a37b2550eded45e46a37515e5182d157

+++ src/ImageProcessorCore/Samplers/Resamplers/BoxResampler.cs
@@ -12,12 +12,12 @@ namespace ImageProcessorCore
     public class BoxResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 0.5d;
+        public float Radius => 0.5F;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            if (x > -0.5 && x <= 0.5)
+            if (x > -0.5F && x <= 0.5F)
             {
                 return 1;
             }
