+++ src/ImageProcessorCore/Samplers/Resamplers/CatmullRomResampler.cs
@@ -14,13 +14,13 @@ namespace ImageProcessorCore
     public class CatmullRomResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            const double B = 0;
-            const double C = 0.5d;
+            const float B = 0;
+            const float C = 0.5F;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
