+++ src/ImageProcessorCore/Samplers/Resamplers/IResampler.cs
@@ -13,15 +13,15 @@ namespace ImageProcessorCore
         /// <summary>
         /// Gets the radius in which to sample pixels.
         /// </summary>
-        double Radius { get; }
+        float Radius { get; }
 
         /// <summary>
         /// Gets the result of the interpolation algorithm.
         /// </summary>
         /// <param name="x">The value to process.</param>
         /// <returns>
-        /// The <see cref="double"/>
+        /// The <see cref="float"/>
         /// </returns>
-        double GetValue(double x);
+        float GetValue(float x);
     }
 }
